[Code review] registration controller dry isssue
